[UPDATE] Filtro implementado por estado da diligencia do projeto
